Revert "more changes"
This reverts commit fe204d9c915c866c1ff3894287e66e9d3494dc62.

+++ code/renderervk/tr_main.cpp
@@ -1234,7 +1234,7 @@ static bool R_MirrorViewBySurface(const drawSurf_t &drawSurf, int entityNum)
 	// create dedicated set for each view
 	if (r_numdlights + oldParms.num_dlights <= arrayLen(backEndData->dlights))
 	{
-		unsigned int i;
+		int i;
 		newParms.dlights = oldParms.dlights + oldParms.num_dlights;
 		newParms.num_dlights = oldParms.num_dlights;
 		r_numdlights += oldParms.num_dlights;
@@ -1620,7 +1620,7 @@ static void R_SortDrawSurfs(drawSurf_t &drawSurfs, int numDrawSurfs)
 		dlight_t *dl;
 		// all the lit surfaces are in a single queue
 		// but each light's surfaces are sorted within its subsection
-		for (i = 0; i < (int)tr.refdef.num_dlights; ++i)
+		for (i = 0; i < tr.refdef.num_dlights; ++i)
 		{
 			dl = &tr.refdef.dlights[i];
 			if (dl->head)
